backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic-settings==2.5.2
xrpl-py==2.6.0
cryptography==43.0.0
python-dotenv==1.0.1
google-generativeai==0.7.2
aiohttp==3.9.1
reportlab==4.0.7
